Align usage of address and ens utilities
