--- src/modules/utils/truncateEthereumAddress.ts
@@ -1,9 +1,0 @@
-const truncateRegex = /^(0x[a-zA-Z0-9]{4})[a-zA-Z0-9]+([a-zA-Z0-9]{4})$/;
-
-export const truncateEthAddress = (address?: string) => {
-    const match = address?.match(truncateRegex);
-    if (!match) {
-        return address;
-    }
-    return `${match[1]}\u2026${match[2]}`;
-};
